+++ projects/ks89/angular-modal-gallery/src/lib/components/modal-gallery/modal-gallery.service.ts
@@ -3,35 +3,28 @@ import { Injectable, Injector, ComponentRef } from '@angular/core';
 import { GlobalPositionStrategy, Overlay, OverlayConfig, OverlayRef } from '@angular/cdk/overlay';
 import { ComponentPortal, PortalInjector } from '@angular/cdk/portal';
 
+import { Subject } from 'rxjs';
+
 import { DIALOG_DATA } from './modal-gallery.tokens';
 import { ModalGalleryComponent } from './modal-gallery.component';
 import { ModalGalleryRef } from './modal-gallery-ref';
 import { Image, ImageModalEvent } from '../../model/image.class';
-import { ConfigService, LibConfig } from '../../services/config.service';
+import { ConfigService } from '../../services/config.service';
 import { ButtonEvent } from '../../model/buttons-config.interface';
-import { Subject } from 'rxjs';
+import { ModalGalleryConfig } from '../../model/modal-gallery-config.interface';
+import { LibConfig } from '../../model/lib-config.interface';
 
-export interface ModalGalleryInput {
-  id: number;
-  images: Image[];
-  currentImage: Image;
-  libConfig?: LibConfig;
+// private interface used only in this file
+interface ModalDialogConfig {
+  panelClass: string;
+  hasBackdrop: boolean;
+  backdropClass: string;
 }
 
-export interface ModalGalleryConfig {
-  panelClass?: string;
-  hasBackdrop?: boolean;
-  backdropClass?: string;
-  config?: ModalGalleryInput;
-  dialogRef?: ModalGalleryRef;
-}
-
-const DEFAULT_CONFIG: ModalGalleryConfig = {
+const DEFAULT_DIALOG_CONFIG: ModalDialogConfig = {
   hasBackdrop: true,
   backdropClass: 'ks-modal-gallery-backdrop',
-  panelClass: 'ks-modal-gallery-panel',
-  config: undefined,
-  dialogRef: undefined
+  panelClass: 'ks-modal-gallery-panel'
 };
 
 @Injectable({ providedIn: 'root' })
@@ -43,15 +36,13 @@ export class ModalGalleryService {
 
   constructor(private injector: Injector, private overlay: Overlay, private configService: ConfigService) {}
 
-  open(config: ModalGalleryConfig = {}): ModalGalleryRef | undefined {
-    // Override default configuration
-    const dialogConfig = { ...DEFAULT_CONFIG, ...config };
+  open(config: ModalGalleryConfig): ModalGalleryRef | undefined {
     // Returns an OverlayRef which is a PortalHost
-    const overlayRef: OverlayRef = this.createOverlay(dialogConfig);
+    const overlayRef: OverlayRef = this.createOverlay();
     // Instantiate remote control
     this.dialogRef = new ModalGalleryRef(overlayRef);
 
-    const overlayComponent: ModalGalleryComponent = this.attachDialogContainer(overlayRef, dialogConfig, this.dialogRef);
+    const overlayComponent: ModalGalleryComponent = this.attachDialogContainer(overlayRef, config, this.dialogRef);
     overlayRef.backdropClick().subscribe(() => {
       if (this.dialogRef) {
         this.dialogRef.closeModal();
@@ -128,8 +119,8 @@ export class ModalGalleryService {
     this.dialogRef.emitButtonAfterHook(event);
   }
 
-  private createOverlay(config: ModalGalleryConfig): OverlayRef {
-    const overlayConfig = this.getOverlayConfig(config);
+  private createOverlay(): OverlayRef {
+    const overlayConfig = this.getOverlayConfig();
     return this.overlay.create(overlayConfig);
   }
 
@@ -142,22 +133,22 @@ export class ModalGalleryService {
     return containerRef.instance;
   }
 
-  private createInjector(overlayConfig: ModalGalleryConfig, dialogRef: ModalGalleryRef): PortalInjector {
+  private createInjector(config: ModalGalleryConfig, dialogRef: ModalGalleryRef): PortalInjector {
     const injectionTokens = new WeakMap();
 
     injectionTokens.set(ModalGalleryRef, dialogRef);
-    injectionTokens.set(DIALOG_DATA, overlayConfig.config);
+    injectionTokens.set(DIALOG_DATA, config);
 
     return new PortalInjector(this.injector, injectionTokens);
   }
 
-  private getOverlayConfig(config: ModalGalleryConfig): OverlayConfig {
+  private getOverlayConfig(): OverlayConfig {
     const positionStrategy: GlobalPositionStrategy = this.overlay.position().global().centerHorizontally().centerVertically();
 
     const overlayConfig: OverlayConfig = new OverlayConfig({
-      hasBackdrop: config.hasBackdrop,
-      backdropClass: config.backdropClass,
-      panelClass: config.panelClass,
+      hasBackdrop: DEFAULT_DIALOG_CONFIG.hasBackdrop,
+      backdropClass: DEFAULT_DIALOG_CONFIG.backdropClass,
+      panelClass: DEFAULT_DIALOG_CONFIG.panelClass,
       scrollStrategy: this.overlay.scrollStrategies.block(),
       positionStrategy
     });