+++ projects/ks89/angular-modal-gallery/src/lib/components/modal-gallery/modal-gallery.tokens.ts
@@ -1,3 +1,4 @@
 import { InjectionToken } from '@angular/core';
+import { ModalGalleryConfig } from '../../model/modal-gallery-config.interface';
 
-export const DIALOG_DATA = new InjectionToken<{}>('DIALOG_DATA');
+export const DIALOG_DATA = new InjectionToken<ModalGalleryConfig>('DIALOG_DATA');