+++ projects/ks89/angular-modal-gallery/src/lib/components/previews/previews.component.ts
@@ -35,8 +35,9 @@ import { SlideConfig } from '../../model/slide-config.interface';
 import { NEXT, PREV } from '../../utils/user-input.util';
 import { getIndex } from '../../utils/image.util';
 import { Action } from '../../model/action.enum';
-import { ConfigService, DEFAULT_PREVIEW_SIZE, LibConfig } from '../../services/config.service';
+import { ConfigService, DEFAULT_PREVIEW_SIZE } from '../../services/config.service';
 import { Size } from '../../model/size.interface';
+import { LibConfig } from '../../model/lib-config.interface';
 
 /**
  * Component with image previews