+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 8.0.0-beta.5
+### Features
+- remove the useless nested 'config' object in ModalGalleryConfig interface.
+
+  To upgrade from all previous 8.0.0 beta releases, change all .open(...) methods passing an object without the nested 'config' object
+
+<br>
+
+**This is a beta version without documentation**
+**If you want to try it, check examples/angular-cli-10 to get tons of working examples**
+
+
 # 8.0.0-beta.4
 ### Features
 - new modal-gallery CSS classes used to set the backdrop **(BREAKING CHANGES)**